Move password related files to the relevant feature
