+++ features/admin.validation.v1/api/update-password-expiry-properties.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -18,9 +18,7 @@
 
 import { AsgardeoSPAClient, HttpRequestConfig } from "@asgardeo/auth-react";
 import { store } from "@wso2is/admin.core.v1";
-import {
-    UpdateMultipleGovernanceConnectorsInterface
-} from "@wso2is/admin.server-configurations.v1";
+import { UpdateGovernanceConnectorConfigInterface } from "@wso2is/admin.server-configurations.v1";
 import { IdentityAppsApiException } from "@wso2is/core/exceptions";
 import { HttpMethods } from "@wso2is/core/models";
 import { AxiosError, AxiosRequestConfig, AxiosResponse } from "axios";
@@ -34,23 +32,34 @@ const httpClient: (
     AsgardeoSPAClient.getInstance()
 );
 
-export const updatePasswordPolicyProperties = (
-    data: UpdateMultipleGovernanceConnectorsInterface
-): Promise<void> => {
+/**
+ * Updates the password expiry properties.
+ *
+ * @param data - The data to update the governance connector configuration.
+ * @returns A promise that resolves with the response data or rejects with an error.
+ *
+ * @throws IdentityAppsApiException If the response status is not 200 or if an error occurs during the request.
+ *
+ * @example
+ * updatePasswordExpiryProperties(data)
+ *  .then(...)
+ *  .catch(...);
+ */
+const updatePasswordExpiryProperties = (data: UpdateGovernanceConnectorConfigInterface): Promise<any> => {
     const requestConfig: AxiosRequestConfig = {
         data,
         headers: {
             "Content-Type": "application/json"
         },
         method: HttpMethods.PATCH,
-        url: store.getState()?.config?.endpoints?.passwordPoliciesUpdate
+        url: store.getState()?.config?.endpoints?.passwordExpiry
     };
 
     return httpClient(requestConfig)
         .then((response: AxiosResponse) => {
             if (response.status !== 200) {
                 throw new IdentityAppsApiException(
-                    "Received an invalid status code while updating the password expiry and history properties.",
+                    "Received an invalid status code while updating the password expiry properties.",
                     null,
                     response.status,
                     response.request,
@@ -59,11 +68,11 @@ export const updatePasswordPolicyProperties = (
                 );
             }
 
-            return Promise.resolve();
+            return Promise.resolve(response.data);
         })
         .catch((error: AxiosError) => {
             throw new IdentityAppsApiException(
-                "An error ocurred while updating the password expiry and history properties.",
+                "An error ocurred while updating the password expiry properties.",
                 error.stack,
                 error.code,
                 error.request,
@@ -72,3 +81,5 @@ export const updatePasswordPolicyProperties = (
             );
         });
 };
+
+export default updatePasswordExpiryProperties;
